Cache the whole replaced file
Previously caching worker separately, which needed parsing to find the position of the worker string.

LCP down from ~2s to ~1.2s
