+++ src/preload/replacementHelpers/cacheReplacement.ts
@@ -1,37 +1,83 @@
+import moduleReplacements from "../moduleReplacements";
 import { applyReplacements } from "./applyReplacement";
-import { Block } from "./parse";
 import { get, set } from "idb-keyval";
+import jsTokens from "js-tokens";
 
 /**
  * Replacements are slow, so we cache the result. We optimize for the common
  * case of having one fixed calculator_desktop loaded several times, then
  * never used again once the next release rolls out.
  */
-export async function applyReplacementsCached(
-  repls: Block[],
-  file: string,
-  key: "worker" | "main"
+export async function fullReplacementCached(
+  calcDesktop: string,
+  workerAppend: string
 ): Promise<string> {
-  const k = "replacement_cached_" + key;
+  const k = "replacement_cached";
   const cached = await get(k);
-  const hashRepls = cyrb53(JSON.stringify(repls));
-  const hashFile = cyrb53(file);
+  const hashRepls = cyrb53(JSON.stringify(moduleReplacements));
+  const hashFile = cyrb53(calcDesktop);
+  const hashAppend = cyrb53(workerAppend);
   if (
     cached !== undefined &&
     cached.hashRepls === hashRepls &&
-    cached.hashFile === hashFile
+    cached.hashFile === hashFile &&
+    cached.hashAppend === hashAppend
   ) {
     // cache hit :)
     return cached.result;
   }
   // cache miss :(
-  const result = applyReplacements(repls, file);
+  const result = fullReplacement(calcDesktop, workerAppend);
   void set(k, {
     hashRepls,
     hashFile,
+    hashAppend,
     result,
   });
   return result;
+}
+
+function fullReplacement(calcDesktop: string, workerAppend: string) {
+  const newCode = applyReplacements(
+    moduleReplacements.filter((r) => !r.workerOnly),
+    calcDesktop
+  );
+  return applyWorkerReplacements(newCode, workerAppend);
+}
+
+function applyWorkerReplacements(src: string, workerAppend: string): string {
+  // Apply replacements to the worker. This could also be done by tweaking the
+  // Worker constructor, but currently all of these replacements could be
+  // performed outside the main page
+  const tokens = Array.from(jsTokens(src));
+  const workerCodeTokens = tokens.filter(
+    (x) =>
+      x.type === "StringLiteral" &&
+      x.value.length > 200000 &&
+      // JS is sure to have &&. Protects against translations getting longer
+      // than the length cutoff, which is intentionally low in case of huge
+      // improvements in minification.
+      x.value.includes("&&")
+  );
+  if (workerCodeTokens.length > 1)
+    throw new Error("More than one worker code found");
+  const wcToken = workerCodeTokens[0];
+  const newWorker = applyReplacements(
+    moduleReplacements.filter((r) => r.workerOnly),
+    // JSON.parse doesn't work because this is a single-quoted string.
+    // js-tokens tokenized this as a string anyway, so it should be
+    // safely eval'able to a string.
+    // eslint-disable-next-line no-eval
+    (0, eval)(wcToken.value) as string
+  );
+  wcToken.value = JSON.stringify(
+    // Place at the beginning of the code for the source mapping to line up
+    // Call at the end of the code to run after modules defined
+    `function loadDesModderWorker(){${workerAppend}\n}` +
+      newWorker +
+      "\nloadDesModderWorker();"
+  );
+  return tokens.map((x) => x.value).join("");
 }
 
 // https://github.com/bryc/code/blob/fed42df9db547493452e32375c93d7854383e480/jshash/experimental/cyrb53.js
